COAExecutor integrated into FederationManager; fixes #78
