+++ cpswt-core/coa/src/main/java/org/cpswt/coa/COAExecutorEventListener.java
@@ -1,4 +1,4 @@
-package org.cpswt.hla;
+package org.cpswt.coa;
 
 /**
  * COAExecutorEventListeren
